+++ cli/cli.go
@@ -33,7 +33,7 @@ import (
 
 var (
 	// Version holds current version of cli-terraform
-	Version = "1.8.0"
+	Version = "1.9.0"
 )
 
 // Run initializes the cli and runs it