+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Features/Enhancements
 
-* CloudWrapper
+* [IMPORTANT] CloudWrapper
   * Added support for `export-cloudwrapper` command which allows export of `akamai_cloudwrapper_configuration` and `akamai_cloudwrapper_activation` resources
 
 * APPSEC
-  * Add import support for custom client sequence
+  * Added import support for custom client sequence
 
 ### Bug fixes